accelerators/sharepoint-connector/pyproject.toml
[project]
name = "sharepoint-connector"
version = "0.1.0"
description = "SharePoint → Azure AI Search push connector via Microsoft Graph API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.16.0",
    "azure-search-documents>=11.4.0",
    "azure-storage-blob>=12.19.0",
    "azure-storage-queue>=12.9.0",
    "azure-data-tables>=12.5.0",
    "azure-ai-documentintelligence>=1.0.0",
    "pyjwt[crypto]>=2.8.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "python-docx>=1.1.0",
    "pymupdf>=1.24.0",
    "python-pptx>=0.6.23",
    "beautifulsoup4>=4.12.0",
    "openpyxl>=3.1.0",
    "extract-msg>=0.48.0",
    "odfpy>=1.4.1",
    "ebooklib>=0.18",
    "striprtf>=0.0.26",
]

[project.optional-dependencies]
func = [
    "azure-functions>=1.17.0",
]
dev = [
    "ruff>=0.3.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"
